// src/types/verification-code.type.ts
export enum VCodeType {
  REGISTER_EMAIL = 'register-email',
  CHANGE_CURRENT_EMAIL = 'change-current-email',
  CHANGE_NEW_EMAIL = 'change-new-email',

  SET_PASSWORD = 'set-password',
  CHANGE_PASSWORD = 'change-password',
  RESET_PASSWORD = 'reset-password',
  
  LOGIN_EMAIL = 'login-email',
}